requirements.txt
aiohttp==3.3.2
aiohttp_jinja==1.0.0
pyaml